Add files via upload
The R code used to run my linear mixed model for Shannon's Index of Alpha Diversity and Hygienic Behaviour Scores. Both models were used with either nested replicates or nested collar in order to explore the entire breadth of the data, but the model was run with combined replicates and collars too (although neither model was significant for either test).